+++ crates/constants/src/chains/mod.rs
@@ -1,0 +1,4 @@
+pub mod pecorino;
+
+#[cfg(any(test, feature = "test-utils"))]
+pub mod test_utils;
